Package Homework01 refactored to Homework00, Homework01 created and Practice 1 made.
