+++ src/main/java/com/nikolenko/homeworks/homework_00/Equations19_26.java
@@ -1,4 +1,4 @@
-package com.nikolenko.homeworks.homework_01;
+package com.nikolenko.homeworks.homework_00;
 
 public class Equations19_26 {
     private KeyBoard keyBoard;
